extracting Authentication interface
