+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/websocket/WebSocketService.java
@@ -18,6 +18,7 @@ import static com.google.common.collect.Streams.stream;
 
 import org.hyperledger.besu.ethereum.api.jsonrpc.authentication.AuthenticationService;
 import org.hyperledger.besu.ethereum.api.jsonrpc.authentication.AuthenticationUtils;
+import org.hyperledger.besu.ethereum.api.jsonrpc.authentication.DefaultAuthenticationService;
 import org.hyperledger.besu.ethereum.api.jsonrpc.websocket.subscription.SubscriptionManager;
 
 import java.net.InetSocketAddress;
@@ -70,7 +71,7 @@ public class WebSocketService {
         vertx,
         configuration,
         websocketRequestHandler,
-        AuthenticationService.create(vertx, configuration));
+        DefaultAuthenticationService.create(vertx, configuration));
   }
 
   private WebSocketService(
@@ -139,10 +140,18 @@ public class WebSocketService {
                       token,
                       user ->
                           websocketRequestHandler.handle(
-                              authenticationService, websocket, buffer.toString(), user, configuration.getRpcApisNoAuth()));
+                              authenticationService,
+                              websocket,
+                              buffer.toString(),
+                              user,
+                              configuration.getRpcApisNoAuth()));
             } else {
               websocketRequestHandler.handle(
-                  Optional.empty(), websocket, buffer.toString(), Optional.empty(), configuration.getRpcApisNoAuth());
+                  Optional.empty(),
+                  websocket,
+                  buffer.toString(),
+                  Optional.empty(),
+                  configuration.getRpcApisNoAuth());
             }
           });
 
@@ -160,10 +169,18 @@ public class WebSocketService {
                       token,
                       user ->
                           websocketRequestHandler.handle(
-                              authenticationService, websocket, payload, user, configuration.getRpcApisNoAuth()));
+                              authenticationService,
+                              websocket,
+                              payload,
+                              user,
+                              configuration.getRpcApisNoAuth()));
             } else {
               websocketRequestHandler.handle(
-                  Optional.empty(), websocket, payload, Optional.empty(), configuration.getRpcApisNoAuth());
+                  Optional.empty(),
+                  websocket,
+                  payload,
+                  Optional.empty(),
+                  configuration.getRpcApisNoAuth());
             }
           });
 
@@ -230,7 +247,7 @@ public class WebSocketService {
       router
           .post("/login")
           .produces(APPLICATION_JSON)
-          .handler(AuthenticationService::handleDisabledLogin);
+          .handler(DefaultAuthenticationService::handleDisabledLogin);
     }
 
     router.route().handler(WebSocketService::handleHttpNotSupported);
